requirements.txt
pandas
numpy
requests
beautifulsoup4
pytz
psycopg2-binary
SQLAlchemy
joblib
scikit-learn
schedule
